Adding my implementation of Connect4. Still WIP
Still WIP, but I suggest you all take a look at my load function, it takes in the connect 4 dataset and produces a 6x7 numpy array of the board and the game state in a last. Goal is to have load return two numpy arrays X and Y. Where X records snapshot of all the game boards and Y is the associated outcome in numpy array format.