made changes to my connect4. I noticed a flaw in my load function.
I have made a few changes to my Connect4. It should be human to human playable but doesn't have a NN yet. I noticed a flaw with my load function, in that while it's nice that it converts the dataset values to numpy array, it simply takes too long process given that there are 67,557 'games' in the dataset