AI Human interaction for connect 4 now working.
I've finished my FFNN and connect 4 implementation. Now you can play a round against my AI as player 1. Keep in mind  the accuracy of my FFNN was  around 63% on training so it sometimes doesn't make the obvious optimal move and you may have to allow it to win at times.